fix(client): address PR review comments
- Change ErrorResponse.statusCode from type: number to type: integer
  in OpenAPI spec so generated models use int instead of float
- Apply max_items truncation in PaginationTransport even when no
  pagination info is detected (single-page responses)
- Improve schema test delay comment to document SCHEMA_TEST_DELAY
  env var and explain the 1.1s default (Katana rate limit)

+++ katana_public_api_client/models/coded_error_response.py
@@ -17,7 +17,7 @@ T = TypeVar("T", bound="CodedErrorResponse")
 class CodedErrorResponse:
     """Error response with an additional application-specific error code for detailed error handling"""
 
-    status_code: float
+    status_code: int
     name: str
     message: str
     code: None | str | Unset = UNSET
